# IADocumentClassifier.API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "IA Document Classifier API",
    "version": "v1"
  },
  "paths": {
    "/api/ClientDocumentType": {
      "get": {
        "tags": [
          "ClientDocumentType"
        ],
        "summary": "Metodo para consultar todos los Clientes por tipos de documentos",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "ClientDocumentType"
        ],
        "summary": "Metodo para Crear Clientes por tipo de documentos",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDocumentTypeDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDocumentTypeDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDocumentTypeDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDocumentTypeDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "ClientDocumentType"
        ],
        "summary": "Metodo para actualizar Clientes por tipo de documentos",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDocumentTypeDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDocumentTypeDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDocumentTypeDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDocumentTypeDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ClientDocumentType/{id}": {
      "get": {
        "tags": [
          "ClientDocumentType"
        ],
        "summary": "Metodo para consultar por Clientes por tipo de documentos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ClientDocumentType"
        ],
        "summary": "Metodo para Eliminar Clientes por tipo de documentos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Clients/{id}": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Metodo para consultar por cliente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Clients"
        ],
        "summary": "Metodo para eliminar clientes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ClientTag": {
      "get": {
        "tags": [
          "ClientTag"
        ],
        "summary": "Metodo para consultar todos los Clientes por Tag",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "ClientTag"
        ],
        "summary": "Metodo para Crear Clientes por Tag",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientTagDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientTagDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientTagDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientTagDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "ClientTag"
        ],
        "summary": "Metodo para actualizar Clientes por Tag",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientTagDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientTagDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientTagDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientTagDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ClientTag/{id}": {
      "get": {
        "tags": [
          "ClientTag"
        ],
        "summary": "Metodo para consultar por Clientes por Tag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ClientTag"
        ],
        "summary": "Metodo para Eliminar Clientes por Tag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Tags/{id}": {
      "get": {
        "tags": [
          "Tags"
        ],
        "summary": "Metodo para consultar especificamente un Tag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Tags"
        ],
        "summary": "Metodo para Borrar Tag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ClientDocumentTypeDTO": {
        "type": "object",
        "properties": {
          "clientDocumentType_Id": {
            "type": "integer",
            "format": "int32"
          },
          "client_Id": {
            "type": "integer",
            "format": "int32"
          },
          "documentType_Id": {
            "type": "integer",
            "format": "int32"
          },
          "path": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {
          "type": "object",
          "additionalProperties": false
        }
      },
...
      "ClientTagDTO": {
        "type": "object",
        "properties": {
          "clientTag_Id": {
            "type": "integer",
            "format": "int32"
          },
          "client_Id": {
            "type": "integer",
            "format": "int32"
          },
          "tag_Id": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}